requirements.txt
Django>=4.2.0
django-cors-headers>=4.3.0
djangorestframework>=3.14.0
asgiref>=3.7.2
psycopg2-binary>=2.9.9
aiohttp>=3.9.1
asyncio>=3.4.3
requests>=2.31.0
statistics>=1.0.3
python-dotenv>=1.0.0
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0